requirements.txt
anime-downloader==5.0.9
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
cfscrape==2.1.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dnspython==2.7.0
fastapi==0.115.6
fuzzywuzzy==0.18.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
outcome==1.3.0.post0
platformdirs==4.3.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.11.1
pyreadline3==3.5.4
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.41.3
tabulate==0.9.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==1.26.6
uvicorn==0.32.1
websocket-client==1.8.0
wsproto==1.2.0
pytest==8.3.3
